feat: implement MapPicker component for location selection in address form
