+++ src/pages/(app)/list-property/steps/building-steps/address.tsx
@@ -38,6 +38,7 @@ import {
   CommandItem,
   CommandList,
 } from "@/components/ui/command";
+import MapPicker from "./google-map";
 
 export default function AddressForm() {
   const dispatch = useDispatch<AppDispatch>();
@@ -58,7 +59,7 @@ export default function AddressForm() {
         zip: true,
         landmark: true,
         area: true,
-        map: true
+        map: true,
       }),
     ),
     defaultValues: {
@@ -150,9 +151,9 @@ export default function AddressForm() {
                     >
                       {field.value?.countryName
                         ? countries.find(
-                          (country) =>
-                            country.countryName === field.value?.countryName,
-                        )?.countryName
+                            (country) =>
+                              country.countryName === field.value?.countryName,
+                          )?.countryName
                         : "Select country..."}
                       <ChevronsUpDown className="size-4 opacity-50" />
                     </div>
@@ -218,9 +219,9 @@ export default function AddressForm() {
                       >
                         {field.value?.stateName
                           ? states.find(
-                            (state) =>
-                              state.stateName === field.value?.stateName,
-                          )?.stateName
+                              (state) =>
+                                state.stateName === field.value?.stateName,
+                            )?.stateName
                           : "Select state..."}
                         <ChevronsUpDown className="size-4 opacity-50" />
                       </div>
@@ -282,8 +283,8 @@ export default function AddressForm() {
                       >
                         {field.value?.cityName
                           ? cities.find(
-                            (city) => city.cityName === field.value?.cityName,
-                          )?.cityName
+                              (city) => city.cityName === field.value?.cityName,
+                            )?.cityName
                           : "Select city..."}
                         <ChevronsUpDown className="size-4 opacity-50" />
                       </div>
@@ -361,22 +362,21 @@ export default function AddressForm() {
           render={({ field }) => (
             <FormItem className="w-full">
               <FormControl>
-                <Input
-                  placeholder="Map"
-                  type="text"
-                  className={cn("text-foreground", {
-                    "border-red-500 focus-visible:ring-red-500":
-                      errors.map?.message,
-                  })}
-                  {...field}
-                />
+                <div className="flex items-center gap-2">
+                  {/* <Input
+                    placeholder="Map"
+                    readOnly
+                    value={field.value}
+                    className={cn("text-foreground", {
+                      "border-red-500 focus-visible:ring-red-500":
+                        errors.map?.message,
+                    })}
+                  /> */}
+                  <MapPicker value={field.value} onChange={field.onChange} />
+                </div>
               </FormControl>
               {errors.map && (
-                <p
-                  className={cn(
-                    "text-sm font-medium text-red-500 dark:text-red-900",
-                  )}
-                >
+                <p className="text-sm font-medium text-red-500 dark:text-red-900">
                   {errors.map?.message}
                 </p>
               )}
